✨ feat: Add PlatformIO CI workflow
This commit adds a new file, `.github/workflows/build-and-release.yml`, which contains the PlatformIO CI workflow. The workflow is triggered on push and pull request events.

The workflow consists of two jobs:

1. **Build**: This job runs on the latest Ubuntu environment and uses the `heltec_wifi_lora_32_V3_HCSR04` environment. It checks out the repository, sets up caching for pip and PlatformIO, installs Python 3.9, installs PlatformIO Core, builds the firmware, and uploads the generated artifacts as build outputs.

2. **Release**: This job runs on the latest Ubuntu environment and depends on the completion of the Build job. It downloads the firmware release artifact, then creates a draft release using the softprops/action-gh-release action. The release draft includes the firmware binaries.

Note: The release job only runs if the event is triggered by a tag that starts with "refs/tags/".